// apps/marketing/src/data/geo/gulf-cities.generated.ts
export const GULF_CITIES: GulfCity[] = [
  { slug: "dubai", name: "Dubai", countrySlug: "uae", countryName: "United Arab Emirates", lat: 25.0772, lng: 55.3093, population: 3790000 },
  { slug: "abu-dhabi", name: "Abu Dhabi", countrySlug: "uae", countryName: "United Arab Emirates", lat: 24.4512, lng: 54.3970, population: 1807000 },
  { slug: "sharjah", name: "Sharjah", countrySlug: "uae", countryName: "United Arab Emirates", lat: 25.3342, lng: 55.4122, population: 1800000 },
  { slug: "ajman", name: "Ajman", countrySlug: "uae", countryName: "United Arab Emirates", lat: 25.4018, lng: 55.4788, population: 490035 },
  { slug: "ras-al-khaimah", name: "Ras Al Khaimah", countrySlug: "uae", countryName: "United Arab Emirates", lat: 25.7895, lng: 55.9432, population: 351943 },
  { slug: "fujairah", name: "Fujairah", countrySlug: "uae", countryName: "United Arab Emirates", lat: 25.1164, lng: 56.3414, population: 118933 },
  { slug: "doha", name: "Doha", countrySlug: "qatar", countryName: "Qatar", lat: 25.2855, lng: 51.5310, population: 344939 },
  { slug: "al-wakrah", name: "Al Wakrah", countrySlug: "qatar", countryName: "Qatar", lat: 25.1715, lng: 51.6034, population: 26436 },
  { slug: "al-khor", name: "Al Khor", countrySlug: "qatar", countryName: "Qatar", lat: 25.6839, lng: 51.5058, population: 18923 },
  { slug: "lusail", name: "Lusail", countrySlug: "qatar", countryName: "Qatar", lat: 25.4175, lng: 51.5075, population: 198600 },
  { slug: "muscat", name: "Muscat", countrySlug: "oman", countryName: "Oman", lat: 23.5841, lng: 58.4078, population: 797000 },
  { slug: "seeb", name: "Seeb", countrySlug: "oman", countryName: "Oman", lat: 23.6703, lng: 58.1891, population: 470878 },
  { slug: "sohar", name: "Sohar", countrySlug: "oman", countryName: "Oman", lat: 24.3474, lng: 56.7094, population: 108274 },
  { slug: "salalah", name: "Salalah", countrySlug: "oman", countryName: "Oman", lat: 17.0150, lng: 54.0924, population: 163140 },
  { slug: "nizwa", name: "Nizwa", countrySlug: "oman", countryName: "Oman", lat: 22.9333, lng: 57.5333, population: 72076 },
  { slug: "sur", name: "Sur", countrySlug: "oman", countryName: "Oman", lat: 22.5667, lng: 59.5289, population: 71152 },
  { slug: "riyadh", name: "Riyadh", countrySlug: "saudi-arabia", countryName: "Saudi Arabia", lat: 24.6877, lng: 46.7219, population: 4205961 },
  { slug: "jeddah", name: "Jeddah", countrySlug: "saudi-arabia", countryName: "Saudi Arabia", lat: 21.4901, lng: 39.1862, population: 4697000 },
  { slug: "dammam", name: "Dammam", countrySlug: "saudi-arabia", countryName: "Saudi Arabia", lat: 26.4344, lng: 50.1033, population: 1252523 },
  { slug: "al-khobar", name: "Al Khobar", countrySlug: "saudi-arabia", countryName: "Saudi Arabia", lat: 26.2794, lng: 50.2083, population: 165799 },
  { slug: "jubail", name: "Jubail", countrySlug: "saudi-arabia", countryName: "Saudi Arabia", lat: 27.0174, lng: 49.6225, population: 237274 },
  { slug: "yanbu", name: "Yanbu", countrySlug: "saudi-arabia", countryName: "Saudi Arabia", lat: 24.0895, lng: 38.0618, population: 200161 },
  { slug: "makkah", name: "Makkah", countrySlug: "saudi-arabia", countryName: "Saudi Arabia", lat: 21.4266, lng: 39.8256, population: 1578722 },
  { slug: "madinah", name: "Madinah", countrySlug: "saudi-arabia", countryName: "Saudi Arabia", lat: 24.4686, lng: 39.6142, population: 1300000 },
  { slug: "kuwait-city", name: "Kuwait City", countrySlug: "kuwait", countryName: "Kuwait", lat: 29.3670, lng: 47.9743, population: 60064 },
  { slug: "farwaniya", name: "Farwaniya", countrySlug: "kuwait", countryName: "Kuwait", lat: 29.2775, lng: 47.9586, population: 86525 },
  { slug: "salmiya", name: "Salmiya", countrySlug: "kuwait", countryName: "Kuwait", lat: 29.3339, lng: 48.0761, population: 147649 },
  { slug: "hawally", name: "Hawally", countrySlug: "kuwait", countryName: "Kuwait", lat: 29.3328, lng: 48.0286, population: 164212 },
  { slug: "mangaf", name: "Mangaf", countrySlug: "kuwait", countryName: "Kuwait", lat: 29.0961, lng: 48.1328, population: 39025 },
  { slug: "manama", name: "Manama", countrySlug: "bahrain", countryName: "Bahrain", lat: 26.2279, lng: 50.5857, population: 147074 },
  { slug: "muharraq", name: "Muharraq", countrySlug: "bahrain", countryName: "Bahrain", lat: 26.2572, lng: 50.6119, population: 176583 },
  { slug: "riffa", name: "Riffa", countrySlug: "bahrain", countryName: "Bahrain", lat: 26.1300, lng: 50.5550, population: 115495 },
];